Added Fee type, and change everywhere. Added changes for deposit_event & related event data
